+++ js/usable.js
@@ -1,0 +1,1 @@
+return reasonFor(arguments[0]);